+++ packages/pluggableWidgets/datagrid-web/typings/global.d.ts
@@ -1,0 +1,21 @@
+interface MXSessionLocale {
+    patterns: {
+        date: string;
+        datetime: string;
+        time: string;
+    };
+}
+
+interface MXGlobalObject {
+    session: {
+        getConfig(): { locale: MXSessionLocale };
+    };
+}
+
+declare global {
+    interface Window {
+        mx?: MXGlobalObject;
+    }
+}
+
+export {};